fix: hide orders on /dashboard/admin
