--- web-marketplace/src/components/organisms/AdminNavigation/AdminNavigation.utils.ts
@@ -1,3 +1,0 @@
-export const isSelected = (path: string, location: string) => {
-  return path === location.substring(location.lastIndexOf('/') + 1);
-};
