+++ web-marketplace/src/components/organisms/RegistryLayout/RegistryLayout.Header.tsx
@@ -1,10 +1,5 @@
 import React, { useMemo } from 'react';
 import { useLocation } from 'react-router-dom';
-import {
-  ApolloClient,
-  NormalizedCacheObject,
-  useApolloClient,
-} from '@apollo/client';
 import { useLingui } from '@lingui/react';
 import Box from '@mui/material/Box';
 import { useTheme } from '@mui/styles';
@@ -18,7 +13,6 @@ import { cn } from 'web-components/src/utils/styles/cn';
 
 import { useAuth } from 'lib/auth/auth';
 import { getPaymentMethodsQuery } from 'lib/queries/react-query/registry-server/getPaymentMethodsQuery/getPaymentMethodsQuery';
-import { getOrdersByBuyerAddressQuery } from 'lib/queries/react-query/registry-server/graphql/indexer/getOrdersByBuyerAddress/getOrdersByBuyerAddress';
 import { useWallet } from 'lib/wallet/wallet';
 
 import { getWalletAddress } from 'pages/Dashboard/Dashboard.utils';
@@ -32,6 +26,7 @@ import { RegistryIconLink, RegistryNavLink } from '../../atoms';
 import { ListProject } from '../ListProject/ListProject';
 import { LoginButton } from '../LoginButton/LoginButton';
 import { useOnProfileClick } from './hooks/useOnProfileClick';
+import { useShowOrders } from './hooks/useShowOrders';
 import {
   getBorderBottom,
   getHeaderColors,
@@ -50,11 +45,9 @@ import { getAddress } from './RegistryLayout.utils';
 const RegistryLayoutHeader: React.FC = () => {
   const { _ } = useLingui();
   const { pathname } = useLocation();
-  const { activeAccount, privActiveAccount, loading } = useAuth();
-  const apolloClient = useApolloClient() as ApolloClient<NormalizedCacheObject>;
+  const { activeAccount, privActiveAccount } = useAuth();
 
-  const { wallet, disconnect, isConnected, loginDisabled, activeWalletAddr } =
-    useWallet();
+  const { wallet, disconnect, isConnected, loginDisabled } = useWallet();
   const { accountOrWallet, noAccountAndNoWallet } = useAuthData();
   const theme = useTheme<Theme>();
   const headerColors = useMemo(() => getHeaderColors(theme), [theme]);
@@ -67,20 +60,7 @@ const RegistryLayoutHeader: React.FC = () => {
   const menuItems = useMemo(() => getMenuItems(pathname, _), [pathname, _]);
   const onProfileClick = useOnProfileClick();
 
-  const fiatOrders = activeAccount?.fiatOrdersByAccountId?.nodes;
-  const { data, isLoading: cryptoOrdersLoading } = useQuery(
-    getOrdersByBuyerAddressQuery({
-      client: apolloClient,
-      enabled: !!activeWalletAddr && !!apolloClient,
-      buyerAddress: activeWalletAddr as string,
-    }),
-  );
-  const cryptoOrders = data?.data?.allOrders?.nodes;
-  const ordersLoading = loading || cryptoOrdersLoading;
-  const showOrders =
-    !ordersLoading &&
-    ((fiatOrders && fiatOrders.length > 0) ||
-      (cryptoOrders && cryptoOrders.length > 0));
+  const showOrders = useShowOrders();
 
   const { data: paymentMethodData } = useQuery(
     getPaymentMethodsQuery({
