+++ web-marketplace/src/lib/i18n/locales/es.po
@@ -22,6 +22,9 @@ msgstr ""
 #: src/components/organisms/CreditActivityTable/CreditActivityTable.tsx:65
 #: src/components/organisms/Documentation/Documentation.constants.ts:11
 #: src/pages/CreditClassDetails/CreditClassDetailsWithContent/CreditClassDetailsWithContent.constants.ts:35
+msgid ""
+msgstr ""
+
 #: src/lib/constants/shared.constants.tsx:96
 msgid "- see less"
 msgstr "- ver menos"
@@ -1782,7 +1785,7 @@ msgstr "Editar filtros"
 msgid "edit profile"
 msgstr "editar perfil"
 
-#: src/components/organisms/AdminNavigation/AdminNavigation.constants.tsx:38
+#: src/components/organisms/AdminNavigation/AdminNavigation.utils.tsx:23
 #: src/components/organisms/RolesForm/components/ProfileModal/ProfileModal.tsx:115
 #: src/pages/Dashboard/Dashboard.constants.tsx:54
 msgid "Edit profile"
@@ -2581,7 +2584,7 @@ msgstr "Debe tener destinatarios"
 #~ msgid "my orders"
 #~ msgstr "Mis ordenes"
 
-#: src/components/organisms/AdminNavigation/AdminNavigation.constants.tsx:18
+#: src/components/organisms/AdminNavigation/AdminNavigation.utils.tsx:46
 #: src/pages/Dashboard/Dashboard.constants.tsx:79
 msgid "My orders"
 msgstr "Mis pedidos"
@@ -2778,7 +2781,7 @@ msgstr "Resumen del pedido"
 msgid "orders"
 msgstr "órdenes"
 
-#: src/components/organisms/AdminNavigation/AdminNavigation.constants.tsx:15
+#: src/components/organisms/AdminNavigation/AdminNavigation.utils.tsx:43
 msgid "Orders"
 msgstr "Pedidos"
 
@@ -3094,7 +3097,7 @@ msgstr "Proceda con precaución: ¡esto no se puede deshacer!"
 msgid "profile"
 msgstr "perfil"
 
-#: src/components/organisms/AdminNavigation/AdminNavigation.constants.tsx:35
+#: src/components/organisms/AdminNavigation/AdminNavigation.utils.tsx:20
 #: src/pages/ProfileEdit/ProfileEdit.constants.tsx:7
 msgid "Profile"
 msgstr "Perfil"
@@ -3777,7 +3780,7 @@ msgstr "Configurar una billetera Keplr →"
 msgid "settings"
 msgstr "ajustes"
 
-#: src/components/organisms/AdminNavigation/AdminNavigation.constants.tsx:43
+#: src/components/organisms/AdminNavigation/AdminNavigation.utils.tsx:28
 #: src/pages/Dashboard/Dashboard.constants.tsx:62
 #: src/pages/Settings/Settings.tsx:52
 msgid "Settings"
